Various cleanup Typo corrections Compile regexes before use Cache result of string manipulation if used more than once Move no-op code in PulseTask to a position in which it makes sense Create abstract superclasses of some similar ICs Un-wrap some lines Add braces where they should be Whoever did Integer.parseInt() in a loop condition of tight nested loops should be ashamed Allow (unused) planter to plant potato, carrot, flowers, and mushrooms Weaken some field access modifiers Change some things to switches where it counts Replace repeated string concatenations with StringBuilders Give utility classes private constructors Allow potatoes to be cooked by cooking pots (Also, for ore smelting - have you considered iterating Bukkit's furnace recipe list?) Removed players from painting editors on quit (to clean up properly) Replaced unnecessary wrapper type (Double) with primitive equivalent

+++ circuits/src/main/java/com/sk89q/craftbook/gates/logic/Delayer.java
@@ -17,13 +17,13 @@ import com.sk89q.craftbook.ic.ICVerificationException;
  */
 public class Delayer extends AbstractIC {
 
-    private int delay = 1;
+    private long delay = 1;
 
     public Delayer(Server server, ChangedSign block, ICFactory factory) {
 
         super(server, block, factory);
         try {
-            delay = Integer.parseInt(getSign().getLine(2));
+            delay = Long.parseLong(getSign().getLine(2));
         } catch (Exception ignored) {
         }
     }
